Creating base class instead used by both platforms.
